requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
idna==2.6
pkg-resources==0.0.0
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
urllib3==1.22
